requirements.txt
certifi==2019.9.11
dal-admin-filters==0.3.3
Django==2.2.7
django-appconf==1.0.3
django-autocomplete-light==3.4.1
django-imagekit==4.0.2
django-leaflet==0.25.0
django-map-widgets==0.2.2
django-material==1.5.8
django-material-admin==1.5.0
django-taggit==1.1.0
GDAL==2.3.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.3
pilkit==2.0
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
